pyproject.toml
[project]
name = "nlp2sql-tuto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "langchain[google-genai]>=1.2.3",
    "langgraph>=1.0.5",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv[cli]>=1.2.1",
    "sqlalchemy>=2.0.46",
    "sqlglot>=28.6.0",
    "uvicorn>=0.40.0",
]
